Add files via upload
Visualization of the loranet. Blue dots are different lora receiver/transmitter devices. Messages can be send and received from these. We have to figure out, the best way to transmitt the messages within the net. There should also be a connection to the other nets of BacNet. We have to figure out, whether we only feed into the other nets or whether and which messages can be received.