demo-script: hot-fix for feed.write bug
The crypto object instantiates the HMAC class differently if it is
created or read from file. If the HMAC object is created HMAC.ss (shared secret)
is a represented as bytes but when passed as a argument the sample
code passes a string.

As an example the feed class expects a string, otherwise it throws this
error:

Traceback (most recent call last):
  File "demo.py", line 65, in <module>
    alice_feed.write(["bacnet/chat", "Hey, it is me Alice!"])
  File "./lib/feed.py", line 69, in write
    self._append(w)
  File "./lib/feed.py", line 51, in _append
    p.open('a')
AttributeError: 'NoneType' object has no attribute 'open'